Remove support for Python 3.2
A recent release of flake8 does not support Python 3.2. This recent
release causes Travis to fail on Python 3.2. Python 3.2 has been EOL
since 2016-02-20. It is no longer receiving bug or security fixes. As it
is EOL, has little usage, and causing test failures, remove it.

Information on supported Python versions:

https://docs.python.org/devguide/index.html#status-of-python-branches