Handle namespace/app_name like the django `include` function
This is done by simply calling `include` and decorating the result,
instead of using our own code - which was wrong - to mimic the `include`
behavior.

The only difference is that we accept "3-tuple" to be able to decorate
stuff like `admin.site.urls`.

This commit also had a real use case test when passing an iterable of
decorators, not only just one, and show an example of this in the
README.

This commit is *breaking* the old (and erroneous) behaviour of the
`decorator_include` function so it will be for the next MAJOR version.